containers/join-server/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
psycopg2-binary>=2.9
drf-spectacular>=0.25
